+++ lact-schema/src/lib.rs
@@ -439,7 +439,7 @@ pub struct DeviceStats {
 pub struct FanStats {
     pub control_enabled: bool,
     pub control_mode: Option<FanControlMode>,
-    pub static_speed: Option<f64>,
+    pub static_speed: Option<f32>,
     pub curve: Option<FanCurveMap>,
     pub pwm_current: Option<u8>,
     pub speed_current: Option<u32>,
@@ -550,7 +550,7 @@ pub struct FanOptions<'a> {
     pub id: &'a str,
     pub enabled: bool,
     pub mode: Option<FanControlMode>,
-    pub static_speed: Option<f64>,
+    pub static_speed: Option<f32>,
     pub curve: Option<FanCurveMap>,
     #[serde(default)]
     pub pmfw: PmfwOptions,
